Add cast.
git-svn-id: https://svn.r-project.org/R/trunk@85497 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/startup.c
@@ -326,7 +326,7 @@ static void SetMaxSize(R_size_t vsize, R_size_t nsize)
 	snprintf(msg, 1024,
 		 "WARNING: too small maximum for language heap (n)size '%lu' ignored,"
 		 " the current usage '%lu' is already larger\n",
-		 (unsigned long) nsize, R_NSize);
+		 (unsigned long) nsize, (unsigned long) R_NSize);
 	R_ShowMessage(msg);
     }
 }
